+++ src/test/e2e-emulated/exit-codes/tact-reserved-contract-errors.spec.ts
@@ -4,6 +4,28 @@ import { Blockchain } from "@ton/sandbox";
 import { ReservedContractErrorsTester as TestContract } from "@/test/e2e-emulated/exit-codes/output/tact-reserved-contract-errors_ReservedContractErrorsTester";
 import "@ton/test-utils";
 
+type ExitCodeValue = 128 | 129 | 130 | 132 | 133 | 134 | 136 | 138;
+type ExitCodeKey =
+    | "TactExitCodeNullReferenceException"
+    | "TactExitCodeInvalidSerializationPrefix"
+    | "TactExitCodeInvalidIncomingMessage"
+    | "TactExitCodeAccessDenied"
+    | "TactExitCodeContractStopped"
+    | "TactExitCodeInvalidArgument"
+    | "TactExitCodeInvalidStandardAddress"
+    | "TactExitCodeNotBasechainAddress";
+
+const TactExitCodes: Map<ExitCodeKey, ExitCodeValue> = new Map([
+    ["TactExitCodeNullReferenceException", 128],
+    ["TactExitCodeInvalidSerializationPrefix", 129],
+    ["TactExitCodeInvalidIncomingMessage", 130],
+    ["TactExitCodeAccessDenied", 132],
+    ["TactExitCodeContractStopped", 133],
+    ["TactExitCodeInvalidArgument", 134],
+    ["TactExitCodeInvalidStandardAddress", 136],
+    ["TactExitCodeNotBasechainAddress", 138],
+]);
+
 describe("Tact-reserved contract errors", () => {
     let blockchain: Blockchain;
     let treasure: SandboxContract<TreasuryContract>;
@@ -33,79 +55,115 @@ describe("Tact-reserved contract errors", () => {
     });
 
     // 128: Null reference exception
-    it("should test exit code 128", async () => {
-        await testReservedExitCode(128, contract, treasure);
+    it("should test exit code TactExitCodeNullReferenceException(128)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeNullReferenceException",
+            contract,
+            treasure,
+        );
     });
 
     // 129: Invalid serialization prefix
-    // NOTE: Reserved, but due to a number of prior checks it cannot be thrown unless one hijacks
-    //       the contract code before deployment and changes the opcodes of the Messages expected
-    //       to be received in the contract
+    it("should test exit code TactExitCodeInvalidSerializationPrefix(129)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeInvalidSerializationPrefix",
+            contract,
+            treasure,
+        );
+    });
 
     // 130: Invalid incoming message
-    it("should test exit code 130", async () => {
-        await testReservedExitCode(130, contract, treasure);
+    it("should test exit code TactExitCodeInvalidIncomingMessage(130)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeInvalidIncomingMessage",
+            contract,
+            treasure,
+        );
     });
 
     // 131: Constraints error
     // NOTE: Reserved, but never thrown anywhere, can't repro
 
     // 132: Access denied
-    it("should test exit code 132", async () => {
-        await testReservedExitCode(132, contract, treasure);
+    it("should test exit code TactExitCodeAccessDenied(132)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeAccessDenied",
+            contract,
+            treasure,
+        );
     });
 
     // 133: Contract stopped
-    it("should test exit code 133", async () => {
-        await testReservedExitCode(133, contract, treasure);
+    it("should test exit code TactExitCodeContractStopped(133)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeContractStopped",
+            contract,
+            treasure,
+        );
     });
 
     // 134: Invalid argument
-    it("should test exit code 134", async () => {
-        await testReservedExitCode(134, contract, treasure);
+    it("should test exit code TactExitCodeInvalidArgument(134)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeInvalidArgument",
+            contract,
+            treasure,
+        );
     });
 
     // 135: Code of a contract was not found
-    // NOTE: Reserved, but one has to replace the contract code to trigger it
+    // NOTE:
+    // If the code of the contract is missing or does not match the one saved
+    // in the TypeScript wrappers, an error with exit code 135 will be thrown
 
     // 136: Invalid address
-    it("should test exit code 136", async () => {
-        await testReservedExitCode(136, contract, treasure);
+    it("should test exit code TactExitCodeInvalidStandardAddress(136)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeInvalidStandardAddress",
+            contract,
+            treasure,
+        );
     });
 
     // 138: Not a basechain address
-    it("should test exit code 138", async () => {
-        await testReservedExitCode(138, contract, treasure);
+    it("should test exit code TactExitCodeNotBasechainAddress(138)", async () => {
+        await testReservedExitCode(
+            "TactExitCodeNotBasechainAddress",
+            contract,
+            treasure,
+        );
     });
 });
 
 async function testReservedExitCode(
-    code: number,
+    codeName: ExitCodeKey,
     contract: SandboxContract<TestContract>,
     treasure: SandboxContract<TreasuryContract>,
 ) {
-    expect(code).toBeGreaterThanOrEqual(128);
-    expect(code).toBeLessThan(256);
-    expect([128, 130, 132, 133, 134, 136, 138]).toContain(code);
-    type testedExitCodes =
-        | "128"
-        | "130"
-        | "132"
-        | "133"
-        | "134"
-        | "136"
-        | "138";
+    const code = TactExitCodes.get(codeName);
+    if (!code) {
+        throw new Error(`No exit code found for ${codeName}`);
+    }
 
     const sendResult = await contract.send(
         treasure.getSender(),
         { value: toNano("10") },
-        code.toString(10) as testedExitCodes,
+        codeName,
     );
 
+    if (codeName === "TactExitCodeInvalidIncomingMessage") {
+        expect(sendResult.transactions).toHaveTransaction({
+            from: contract.address,
+            to: contract.address,
+            success: false,
+            exitCode: code,
+        });
+        return;
+    }
+
     expect(sendResult.transactions).toHaveTransaction({
-        from: code === 130 ? contract.address : treasure.address,
+        from: treasure.address,
         to: contract.address,
-        success: false,
-        exitCode: code,
+        success: true,
     });
 }